+++ lib/acts_as_moonable.rb
@@ -1,6 +1,7 @@
 require "acts_as_moonable/version"
 require "acts_as_moonable/railtie"
 require "acts_as_moonable/core_ext"
+require "acts_as_moonable/humanable"
 
 module ActsAsMoonable
   # Your code goes here...