+++ test/dummy/app/models/post.rb
@@ -1,0 +1,9 @@
+class Post < ApplicationRecord
+  include ActsAsMoonable::Humanable
+
+  # constants
+  CURRENCY = %w( CNY USD EUR GBP JPY )
+
+  # enums
+  enum status: { edited: 0, published: 1, approved: 2, trashed: 3 }
+end