+++ CHANGELOG.md
@@ -1,5 +1,22 @@
-### v0.1.0 (2021-08-09)
+# ActsAsMoonable
 
-#### 初始发布
+## 待发布
+
+### Features
+* Interval Module
+
+### Bugfixes
+* 无
+
+
+## v0.2.0 (2021-08-09)
+
+### Features
+* Humanable Module
+
+
+## v0.1.0 (2021-08-09)
+
+### 初始发布
 
 * CoreExt Module